build a simple website for a fictional, mobile application called Gamma which is a photo sharing app
Basically, I have a landing page here, a bootstrap formatted page. I
have the navbar, we have the main showcase area or jumbotron as it’s
called in Bootstrap. And then, if you scroll down, I have there three
columns here with images, texts and buttons. And then, I have kind of a
features section. It has a background color. I have an image, have some
bullet points that are using the icons that Bootstrap offers. It’s
responsible, you can see that it changes to fit the screen. You’ll
notice, if we go past a certain point, that the large iPhone image
disappears just because it’s much too big and cluttered on a smaller
screen. You can see as we go down to the mobile size, it looks really
good. So everything fits nicely. And then I just have a basic footer at
the bottom. It’s the home page. I also have a couple other page, the
about page, I have a title heading area, have some text, we’re using
the Accordion module that bootstrap offers. And then, I just have a
simple sidebar with an image and some text. And then, the contact page,
I just have a bootstrap formatted form and sidebar.